Mesh Class with Collider Info
-Added overloaded ImportMesh function to AssetManager
-Updated Mesh class to hold collider info for each Mesh
-Updated Enemy constructor to be more conscise
-Fixed minor bug where moving enemies didn't shoot form their current position
-Trimmed down to a single GameObject constructor, which should correctly handle all cases of GameObject creation
-Fixed type conversion warning in OctreeNode
-Updated Projectile Constructor
-Updated atoi with _atoi_s, which is both safe and not deprecated

+++ DX11Starter/OctreeNode.cpp
@@ -24,7 +24,7 @@ GameObject * OctreeNode::GetObjects()
 
 unsigned int OctreeNode::GetObjectCount()
 {
-	return objects.size();
+	return (unsigned int)objects.size();
 }
 
 void OctreeNode::AddObject(GameObject * obj)
